+++ src/rowgroup.js
@@ -104,32 +104,14 @@ export function readRowGroup(options, { metadata }, groupPlan) {
 }
 
 /**
- * @overload
  * @param {AsyncRowGroup} asyncGroup
  * @param {number} selectStart
  * @param {number} selectEnd
  * @param {string[] | undefined} columns
- * @param {'object'} rowFormat
  * @returns {Promise<Record<string, any>[]>} resolves to row data
  */
-/**
- * @overload
- * @param {AsyncRowGroup} asyncGroup
- * @param {number} selectStart
- * @param {number} selectEnd
- * @param {string[] | undefined} columns
- * @param {'array'} [rowFormat]
- * @returns {Promise<any[][]>} resolves to row data
- */
-/**
- * @param {AsyncRowGroup} asyncGroup
- * @param {number} selectStart
- * @param {number} selectEnd
- * @param {string[] | undefined} columns
- * @param {'object' | 'array'} [rowFormat]
- * @returns {Promise<Record<string, any>[] | any[][]>} resolves to row data
- */
-export async function asyncGroupToRows({ asyncColumns }, selectStart, selectEnd, columns, rowFormat) {
+export async function asyncGroupToRows({ asyncColumns }, selectStart, selectEnd, columns) {
+  // columnData[i] for asyncColumns[i]
   // TODO: do it without flatten
   const asyncPages = await Promise.all(asyncColumns.map(async ({ data }) => {
     const pages = await data
@@ -139,44 +121,23 @@ export async function asyncGroupToRows({ asyncColumns }, selectStart, selectEnd,
     }
   }))
 
-  // careful mapping of column order for rowFormat: array
-  const includedColumnNames = asyncColumns
-    .map(child => child.pathInSchema[0])
-    .filter(name => !columns || columns.includes(name))
-  const columnOrder = columns ?? includedColumnNames
-  const columnIndexes = columnOrder.map(name => asyncColumns.findIndex(column => column.pathInSchema[0] === name))
+  // filter columns
+  const filteredColumns = columns
+    ? asyncColumns.filter(column => columns.includes(column.pathInSchema[0]))
+    : asyncColumns
 
   // transpose columns into rows
   const selectCount = selectEnd - selectStart
-  if (rowFormat === 'object') {
-    /** @type {Record<string, any>[]} */
-    const groupData = Array(selectCount)
-    for (let selectRow = 0; selectRow < selectCount; selectRow++) {
-      const row = selectStart + selectRow
-      // return each row as an object
-      /** @type {Record<string, any>} */
-      const rowData = {}
-      for (let i = 0; i < asyncColumns.length; i++) {
-        const { data, pageSkip } = asyncPages[i]
-        rowData[asyncColumns[i].pathInSchema[0]] = data[row - pageSkip]
-      }
-      groupData[selectRow] = rowData
-    }
-    return groupData
-  }
-
-  /** @type {any[][]} */
-  const groupData = Array(selectCount)
+  /** @type {Record<string, any>[]} */
+  const groupData = new Array(selectCount)
   for (let selectRow = 0; selectRow < selectCount; selectRow++) {
     const row = selectStart + selectRow
-    // return each row as an array
-    const rowData = Array(asyncColumns.length)
-    for (let i = 0; i < columnOrder.length; i++) {
-      const colIdx = columnIndexes[i]
-      if (colIdx >= 0) {
-        const { data, pageSkip } = asyncPages[colIdx]
-        rowData[i] = data[row - pageSkip]
-      }
+    /** @type {Record<string, any>} */
+    const rowData = {}
+    for (let i = 0; i < filteredColumns.length; i++) {
+      const columnIndex = asyncColumns.indexOf(filteredColumns[i])
+      const { data, pageSkip } = asyncPages[columnIndex]
+      rowData[filteredColumns[i].pathInSchema[0]] = data[row - pageSkip]
     }
     groupData[selectRow] = rowData
   }