+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [2.0.0]
+ - Rename `parquetMetadataAsync` to `parquetMetadata`
+ - Previous `parquetMetadata` is now `import { parquetMetadataSync } from 'hyparquet/src/metadata.js'`
+ - Rename `onPage.columnName: string` to `onPage.pathInSchema: string[]`
+ - Remove `rowFormat` and always return rows as objects
+ - Change `byteLengthFromUrl` to accept options object
+ - Move `file` parameter to options object in `parquetMetadata`
+
 ## [1.23.3]
  - Fix `readColumn` truncation in struct columns (#148)